chore(wip): bill ui reminders
